+++ qt/gen_qcolumnview.go
@@ -161,7 +161,7 @@ func (this *QColumnView) SetPreviewWidget(widget *QWidget) {
 
 func (this *QColumnView) SetColumnWidths(list []int) {
 	// For the C ABI, malloc a C array of raw pointers
-	list_CArray := (*[0xffff]C.int)(C.malloc(C.ulong(8 * len(list))))
+	list_CArray := (*[0xffff]C.int)(C.malloc(C.size_t(8 * len(list))))
 	defer C.free(unsafe.Pointer(list_CArray))
 	for i := range list {
 		list_CArray[i] = (C.int)(list[i])
